Mode 'GpuStat' and 'GpuMode' into their own modules
