+++ src/gameboy/gfx/mod.rs
@@ -2,7 +2,11 @@
 mod color;
 mod frame;
 mod gpu;
+mod mode;
+mod stat;
 
 pub use self::color::Color;
 pub use self::gpu::Gpu;
 pub use self::frame::Frame;
+pub use self::mode::GpuMode;
+pub use self::stat::GpuStat;
